[mms-engine] Fix incorrect parameter in mms_engine_handle_cancel
The cancel method should take a string id, but currently a few functions
are still defined with integer ids, including sendMessage and cancel.
sendMessage correctly converts the integer to a string, but cancel had a
bad cast that resulted in a crash when called with an integer id.

We should fix the parameter types on these methods at some point, which
will require a coordinated change with the MmsHandler. Until then, this
prevents the crash and makes the cancel method work as expected.