[mms-dump] Added option to dump attachment data
Also, ignore failure to parse part headers. If we know the total size of the
part headers, such a failure doesn't stop us from continuing to parse the rest
of the PDU. The purpose of this tool is not to validate the PDU but to dump as
much information as possible even if PDU is not perfectly valid.