[mms-settings] Changed dconf path for MMS settings
Previously, MMS settings were stored right in the IMSI directory at the
root of the dconf namespace. They have been moved to /imsi/ folder and
MMS settings now live in the mms subfolder, i.e. /IMSI/ -> /imsi/IMSI/mms/

Old settings are migrated to the new location the first time they are
requested. Instead of GSettings we now directly use DConf because it
appears to be impossible to properly implement migration with GSettings.
That makes gschema obsolete but we need to keep it around for backward
compatibility reasons.