[mms-lib] Don't delete m-notification.ind if we are keeping temp files
Additionally, move it from the directory with the fake id to the directory
associated with the message once we know which id commhistoryd has assigned
to this message.

These changes are only useful for debugging, if mms-engine is started
with -t option. They have no effect on the normal operation.